Utilize architectury plugin for platform code
