+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/inline/README.md
