+++ velocity/src/main/java/io/github/kosmx/emotes/velocity/executor/VelocityInstance.java
@@ -23,21 +23,6 @@ public class VelocityInstance extends EmoteInstance {
     }
 
     @Override
-    public IDefaultTypes getDefaults() {
-        return null;
-    }
-
-    @Override
-    public IGetters getGetters() {
-        return null;
-    }
-
-    @Override
-    public IClientMethods getClientMethods() {
-        return null;
-    }
-
-    @Override
     public boolean isClient() {
         return false;
     }
